backend/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
spacy==3.7.2
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
tika==2.6.0
transformers==4.35.2
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2